python/requirements.txt
accelerate==1.5.2
acres==0.3.0
albucore==0.0.23
albumentations==2.0.5
altgraph==0.17.4
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.8.0
apispec==6.6.1
apispec-webframeworks==1.1.0
arrow==1.3.0
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
attrs==25.3.0
audioread==3.0.1
Automat==24.8.1
awscli==1.38.23
backcall==0.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bidict==0.23.1
binaryornot==0.4.4
blinker==1.9.0
boto3==1.37.23
botocore==1.37.23
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
ci-info==0.3.0
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
constantly==23.10.4
contourpy==1.3.1
cookiecutter==2.6.0
cycler==0.12.1
Cython==3.0.12
debugpy==1.8.6
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docopt==0.6.2
docutils==0.16
docx==0.2.4
durationpy==0.9
easyocr==1.7.2
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing==2.1.0
faiss-cpu==1.10.0
fastapi==0.115.12
filelock==3.18.0
fire==0.7.0
fitz==0.0.1.dev2
Flask==3.1.0
flask-cors==5.0.1
Flask-RESTful==0.3.10
Flask-SocketIO==5.4.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
ftfy==6.3.1
gevent==24.11.1
gevent-websocket==0.10.1
gitignore_parser==0.1.11
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
importlib_resources==6.5.2
incremental==24.7.2
ipykernel==6.29.5
ipython==8.12.3
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jedi==0.19.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.7.4.5
keyring==8.7
keyrings.alt==5.0.2
kiwisolver==1.4.8
kthread==0.2.3
kubernetes==32.0.1
langchain==0.3.21
langchain-core==0.3.48
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.7
langsmith==0.3.18
lazy_loader==0.4
librosa==0.11.0
llama_cpp_python==0.3.7
llvmlite==0.44.0
lmdb==1.6.2
looseversion==1.3.0
lxml==5.3.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.21.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.9.2
numba==0.61.2
numpy==2.1.1
oauthlib==3.2.2
onnxruntime==1.21.0
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.2
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-asgi==0.52b0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-util-http==0.52b0
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.10.15
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
paddleocr==2.10.0
paddlepaddle==3.0.0
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
pefile==2023.2.7
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
pooch==1.8.2
posthog==3.21.0
prompt_toolkit==3.0.48
protobuf==5.29.3
prov==2.0.1
psutil==6.0.0
pure_eval==0.2.3
puremagic==1.28
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
pydub==0.25.1
Pygments==2.18.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pymongo==4.7.2
PyMuPDF==1.25.4
pyparsing==3.2.1
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-bidi==0.6.6
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.12.1
pytube==15.0.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.3
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.12.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.7.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.11.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==3.4.1
setuptools==75.8.2
shapely==2.0.7
shellingham==1.5.4
simple-websocket==1.0.0
simplejson==3.20.1
simsimd==6.2.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.46.2
stringzilla==3.12.3
sympy==1.13.1
tabulate==0.9.0
taipy==4.0.2
taipy-common==4.0.2
taipy-core==4.0.2
taipy-gui==4.0.2
taipy-rest==4.0.2
taipy-templates==4.0.2
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.3.13
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.50.1
Twisted==24.7.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata==2025.1
tzlocal==5.2
ultralytics==8.3.98
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
voluptuous==0.15.2
watchdog==4.0.1
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
yarg==0.1.9
yolo==0.3.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
